% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kirpPheno_df.R
\docType{data}
\name{kirpPheno_df}
\alias{kirpPheno_df}
\title{Kidney Papillary Cell Carcinoma (KIRP) Clinical Data}
\format{A data frame containing TCGA sample IDs (\code{Sample}), overall
  survival time in days (\code{time}), death indicator (\code{status}), and
  male-sex indicator for 349 subjects.}
\source{
Clinical KIRP data from the TCGA hub as hosteb by Xena:
  \url{https://xenabrowser.net/hub/}

  Cleaning script: \code{inst/scripts/download_KIRP.R}
}
\usage{
kirpPheno_df
}
\description{
Clinical data from The Cancer Genome Atlas' KIRP cohort
}
\keyword{datasets}
