% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kirpRNAseq_aespcOut.R
\docType{data}
\name{kirpRNAseq_aespcOut}
\alias{kirpRNAseq_aespcOut}
\title{Kidney Papillary Cell Carcinoma (KIRP) RNAseq AES-PCA Output}
\format{An object of class \code{aespcOut}. Access the principal components
  and corresponding loadings for each pathway from this analysis with the
  function \code{\link[pathwayPCA]{getPathPCLs}}. Access the table of pathway
  significance levels with the function
  \code{\link[pathwayPCA]{getPathpVals}}.}
\source{
calculated via the package vignette
}
\usage{
kirpRNAseq_aespcOut
}
\description{
The AES-PCA analysis of the TCGA KIRP data set included in this
  package
}
\keyword{datasets}
