% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ovCNV_aespcOut.R
\docType{data}
\name{ovCNV_aespcOut}
\alias{ovCNV_aespcOut}
\title{Ovarian Cancer Copy-number Variation AES-PCA Output}
\format{An object of class \code{aespcOut}. Access the principal components
  and corresponding loadings for each pathway from this analysis with the
  function \code{\link[pathwayPCA]{getPathPCLs}}. Access the table of pathway
  significance levels with the function
  \code{\link[pathwayPCA]{getPathpVals}}.}
\source{
calculated via the package vignette
}
\usage{
ovCNV_aespcOut
}
\description{
The AES-PCA analysis of the TCGA ovarian cancer data set
  included in this package
}
\keyword{datasets}
