% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ovPheno_df.R
\docType{data}
\name{ovPheno_df}
\alias{ovPheno_df}
\title{Ovarian Cancer Clinical Data}
\format{A data frame containing TCGA sample IDs (\code{Sample}), overall
  survival time in days (\code{OS_time}), and death indicator
  (\code{OS_status}) copy-number values for 579 samples.}
\source{
Clinical
  \url{http://linkedomics.org/data_download/TCGA-OV/}

  Cleaning script: \code{inst/scripts/clean_multi_omics.R}
}
\usage{
ovPheno_df
}
\description{
Clinical Proteomic Tumor Analysis Consortium's the survival
  information from the clinical annotation data.
}
\keyword{datasets}
